attempt to replace service-worker build process with esbuild
