+++ packages/client/package.json
@@ -3,12 +3,13 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "build": "tsc && vite build",
-    "build:dev": "cross-env BUILD_MODE=development vite build",
+    "build": "tsc && yarn sw:build && vite build",
+    "build:dev": "vite --mode=development build",
     "dev": "vite",
-    "dev:vite": "vite",
-    "dev:sw": "tsc --watch --project src/service-worker/worker/tsconfig.sw.json",
-    "dev:with-sw": "run-p --print-label dev:sw dev:vite",
+    "dev:with-sw": "run-p --print-label sw:dev dev",
+    "sw:build": "tsc --noEmit --project src/service-worker/worker/tsconfig.sw.json && yarn sw:transpile",
+    "sw:transpile": "node ./scripts/transpile-sw.cjs",
+    "sw:dev": "yarn sw:transpile --mode=development --watch",
     "format": "prettier --write .",
     "lint": "eslint .",
     "preview": "vite preview",
@@ -44,6 +45,7 @@
     "jest": "^28",
     "jest-environment-jsdom": "^29.0.1",
     "lefthook": "^1.1.1",
+    "minimist": "1.2.8",
     "npm-run-all": "4.1.5",
     "postcss": "^8.4.21",
     "postcss-cli": "^10.1.0",
@@ -53,7 +55,6 @@
     "postcss-nesting": "^11.2.2",
     "postcss-preset-env": "^8.1.0",
     "prettier": "^2.7.1",
-    "rollup-plugin-typescript": "1.0.1",
     "ts-jest": "^28.0.8",
     "typescript": "^4.8.2",
     "vite": "4.2.1"
