+++ packages/client/public/offline.html
@@ -4,8 +4,11 @@
   <meta charset="UTF-8">
   <title>Connection Error</title>
 <body>
-<main style="display: flex; align-items: center; justify-content: space-evenly;
-  height: 100vh; min-width: 100vw;">
+<main style="
+  display: flex; align-items: center; justify-content: space-evenly;
+  height: 100vh; min-width: 100vw; 
+  color: #fafafa; background-color: #010318;
+  ">
   <h1>Похоже, ваше приложение не в сети</h1>
 </main>
 </body>
