+++ packages/client/src/service-worker/worker/helpers.ts
@@ -1,0 +1,7 @@
+export function getURL(request: RequestInfo): string {
+  if (typeof request === 'string') {
+    return request;
+  } else {
+    return new URL(request.url).pathname;
+  }
+}
