+++ packages/client/src/service-worker/worker/service-worker.ts
@@ -1,76 +1,5 @@
-type CacheHandler = {
-  addResourcesToCache: (resources: RequestInfo[]) => Promise<void>;
-  putInCache: ({
-    request,
-    response,
-  }: {
-    request: RequestInfo;
-    response: Response;
-  }) => Promise<void>;
-  deleteCache: (cacheName: string) => Promise<void>;
-  deleteOldCaches: ({ keepList }?: { keepList: string[] }) => Promise<void>;
-};
-
-const getURL = (request: RequestInfo): string => {
-  if (typeof request === 'string') {
-    return request;
-  } else {
-    return new URL(request.url).pathname;
-  }
-};
-
-function makeCacheHandler(cacheName: string): CacheHandler {
-  const addResourcesToCache: CacheHandler['addResourcesToCache'] =
-    async resources => {
-      const cache = await caches.open(cacheName);
-      await cache.addAll(resources);
-    };
-
-  const putInCache: CacheHandler['putInCache'] = async ({
-    request,
-    response,
-  }) => {
-    const cache = await caches.open(cacheName);
-    let logString = `NETWORK REQUEST: ${getURL(request)}; STATUS: ${
-      response.status
-    }\n`;
-    logString += `PUT REQUEST '${getURL(request)}' TO CACHE ${cacheName}`;
-
-    try {
-      await cache.put(request, response);
-    } catch (error) {
-      logString += `FAILED WITH ERROR: ${error}`;
-    }
-
-    console.log(logString);
-  };
-
-  const deleteCache: CacheHandler['deleteCache'] = async (
-    cacheToDelete: string
-  ) => {
-    console.log(`DELETE CACHE ${cacheToDelete}`);
-    await caches.delete(cacheToDelete);
-  };
-
-  const deleteOldCaches: CacheHandler['deleteOldCaches'] = async (
-    { keepList } = { keepList: [cacheName] }
-  ) => {
-    const cachesKeysList = await caches.keys();
-    console.log(`CASHES: ${JSON.stringify(cachesKeysList)}`);
-    const cachesToDelete = cachesKeysList.filter(
-      key => !keepList.includes(key)
-    );
-    console.log(`DELETE OLD CACHES ${JSON.stringify(cachesToDelete)}`);
-    await Promise.all(cachesToDelete.map(deleteCache));
-  };
-
-  return {
-    addResourcesToCache,
-    putInCache,
-    deleteCache,
-    deleteOldCaches,
-  };
-}
+import { makeCacheHandler } from './cache-handler';
+import { getURL } from './helpers';
 
 const sw: ServiceWorkerGlobalScope = self as any;
