+++ packages/client/src/service-worker/worker/tsconfig.sw.json
@@ -5,8 +5,7 @@
     "isolatedModules": false,
     "lib": ["ESNext", "WebWorker"],
     "noEmit": false,
-    "outDir": "../../../public",
-    "sourceMap": true
+    "outDir": "../../../public"
   },
   "include": ["."],
   "exclude": []
