+++ packages/client/vite.config.ts
@@ -1,8 +1,6 @@
 import react from '@vitejs/plugin-react';
 import dotenv from 'dotenv';
 import path from 'path';
-import { InputOptions, OutputOptions, rollup } from 'rollup';
-import rollupPluginTypescript from 'rollup-plugin-typescript';
 import { defineConfig } from 'vite';
 dotenv.config();
 
@@ -19,24 +17,6 @@ if (IS_DEV) {
 }
 console.log(outputNamesOptions);
 
-const CompileTsServiceWorker = () => ({
-  name: 'compile-typescript-service-worker',
-  async writeBundle() {
-    const inputOptions: InputOptions = {
-      input: 'src/service-worker/worker/service-worker.ts',
-      plugins: [rollupPluginTypescript()],
-    };
-    const outputOptions: OutputOptions = {
-      file: `${OUT_DIR}/service-worker.js`,
-      format: 'es',
-    };
-    const bundle = await rollup(inputOptions);
-    await bundle.write(outputOptions);
-    await bundle.close();
-  },
-});
-
-// https://vitejs.dev/config/
 export default defineConfig({
   server: {
     port: Number(process.env.CLIENT_PORT) || 3000,
@@ -44,7 +24,7 @@ export default defineConfig({
   define: {
     __SERVER_PORT__: process.env.SERVER_PORT,
   },
-  plugins: [react(), CompileTsServiceWorker()],
+  plugins: [react()],
   build: {
     outDir: OUT_DIR,
     cssMinify: false,
